+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 1.1 branch
 ### 1.1.11
 * added `TestCase` class;
+* added `objects_map()` global function;
 * `assertFileExtension()` and `assertFileMime()` assert methods can take string
     or an array or a `Traversable` of files;
 * all `ReflectionTrait` methods are now protected. The `setProperty()` method